move search method to search.rs
